+++ tests/test_awstaga.py
@@ -38,7 +38,7 @@ class TestAwstaga(unittest.TestCase):
         mock_tagset.get_tags.return_value = [mock_tagset_tag]
         mock_resource.get_tags.return_value = [mock_resource_tag]
 
-        apply(conf_file='awstaga.yaml', dry_run=False, batch_size=100)
+        apply(conf_file='awstaga.yaml', dry_run=False, batch_size=20)
 
         self.assertEqual(mock_logger.info.call_count, 3)
         mock_logger.info.assert_has_calls([
@@ -85,7 +85,7 @@ class TestAwstaga(unittest.TestCase):
         mock_tagset.get_tags.return_value = [mock_tagset_tag]
         mock_resource.get_tags.return_value = [mock_resource_tag]
 
-        apply(conf_file='awstaga.yaml', dry_run=True, batch_size=100)
+        apply(conf_file='awstaga.yaml', dry_run=True, batch_size=20)
 
         self.assertEqual(mock_logger.info.call_count, 2)
         mock_logger.info.assert_has_calls([